split into regular and path-aware repeat
